GBE: change GEN binary format.
Change all size_t to uint32_t when generate GEN binary, size_t is
non compat between 32bit system and 64bit system, and uint32_t is
enough.

+++ backend/src/backend/gen_program.cpp
@@ -80,7 +80,7 @@ namespace gbe {
     insns = (GenInstruction *)ins;
     insnNum = size / sizeof(GenInstruction);
   }
-  size_t GenKernel::getCodeSize(void) const { return insnNum * sizeof(GenInstruction); }
+  uint32_t GenKernel::getCodeSize(void) const { return insnNum * sizeof(GenInstruction); }
 
   void GenKernel::printStatus(int indent, std::ostream& outs) {
 #ifdef GBE_COMPILER_AVAILABLE
